@@ -1,12 +1,11 @@
 using Strategy.DuckHFDF.Strategies.Abstractions;
 
-namespace Strategy.DuckHFDF.Strategies
+namespace Strategy.DuckHFDF.Strategies;
+
+public class Squeak : IQuackBehavior
 {
-    public class Squeak : IQuackBehavior
+    public void Quack()
     {
-        public void Quack()
-        {
-            Console.WriteLine("Squeak");
-        }
+        Console.WriteLine("Squeak");
     }
 }
