@@ -2,27 +2,25 @@
 
 namespace Observer.WeatherData.Observers;
 
-public class StatisticsDisplay : IObserver, IDisplayElement
+public class CurrentConditionsDisplay : IObserver, IDisplayElement
 {
+    public CurrentConditionsDisplay(ISubject weatherData)
+    {
+        weatherData.RegisterObserver(this);
+    }
+
     public double Temperature { get; set; }
     public double Humidity { get; set; }
-    public double Pressure { get; set; }
 
-    public StatisticsDisplay(ISubject subject)
+    public void Display()
     {
-        subject.RegisterObserver(this);
+        Console.WriteLine($"Current Conditions Display: {Temperature} degrees and {Humidity} % humidity");
     }
 
     public void Update(double temperature, double humidity, double pressure)
     {
         Temperature = temperature;
         Humidity = humidity;
-        Pressure = pressure;
         Display();
-    }
-
-    public void Display()
-    {
-        Console.WriteLine($"StatisticsDisplay: {Temperature}, {Humidity}, {Pressure}");
     }
 }
