@@ -2,25 +2,27 @@
 
 namespace Observer.WeatherData.Observers;
 
-public class CurrentConditionsDisplay : IObserver, IDisplayElement
+public class StatisticsDisplay : IObserver, IDisplayElement
 {
+    public StatisticsDisplay(ISubject subject)
+    {
+        subject.RegisterObserver(this);
+    }
+
     public double Temperature { get; set; }
     public double Humidity { get; set; }
+    public double Pressure { get; set; }
 
-    public CurrentConditionsDisplay(ISubject weatherData)
+    public void Display()
     {
-        weatherData.RegisterObserver(this);
+        Console.WriteLine($"StatisticsDisplay: {Temperature}, {Humidity}, {Pressure}");
     }
-    
+
     public void Update(double temperature, double humidity, double pressure)
     {
         Temperature = temperature;
         Humidity = humidity;
+        Pressure = pressure;
         Display();
-    }
-
-    public void Display()
-    {
-        Console.WriteLine($"Current Conditions Display: {Temperature} degrees and {Humidity} % humidity");
     }
 }
