@@ -1,8 +1,7 @@
-﻿using Observer.WeatherDataObservable.Observers;
-using Observer.WeatherDataObservable.Subjects;
+﻿using Observer.WeatherData.Observers;
+using Observer.WeatherData.Subject;
 
 var weatherData = new WeatherData();
-
 var currentDisplay = new CurrentConditionsDisplay(weatherData);
 var statisticsDisplay = new StatisticsDisplay(weatherData);
 var forecastDisplay = new ForecastDisplay(weatherData);
@@ -10,9 +9,7 @@ var heatIndexDisplay = new HeatIndexDisplay(weatherData);
 
 weatherData.SetMeasurements(80, 65, 30.4d);
 Console.WriteLine();
-
 weatherData.SetMeasurements(82, 70, 29.2d);
 Console.WriteLine();
-
 weatherData.SetMeasurements(78, 90, 29.2d);
 Console.WriteLine();
