@@ -4,13 +4,30 @@ namespace Observer.WeatherDataObservable.Observers;
 
 public class ForecastDisplay : IObserver<WeatherData>, IDisplayElement
 {
+    private readonly IDisposable _unsubscriber;
     private double _currentPressure;
     private double _lastPressure;
-    private readonly IDisposable _unsubscriber;
 
     public ForecastDisplay(IObservable<WeatherData> weatherData)
     {
         _unsubscriber = weatherData.Subscribe(this);
+    }
+
+    public void Display()
+    {
+        Console.Write("Forecast: ");
+        if (_currentPressure > _lastPressure)
+        {
+            Console.WriteLine("Improving weather on the way!");
+        }
+        else if (Math.Abs(_currentPressure - _lastPressure) < 0.01f)
+        {
+            Console.WriteLine("More of the same");
+        }
+        else if (_currentPressure < _lastPressure)
+        {
+            Console.WriteLine("Watch out for cooler, rainy weather");
+        }
     }
 
     public void OnCompleted()
@@ -30,22 +47,5 @@ public class ForecastDisplay : IObserver<WeatherData>, IDisplayElement
         _currentPressure = value.Pressure;
 
         Display();
-    }
-
-    public void Display()
-    {
-        Console.Write("Forecast: ");
-        if (_currentPressure > _lastPressure)
-        {
-            Console.WriteLine("Improving weather on the way!");
-        }
-        else if (Math.Abs(_currentPressure - _lastPressure) < 0.01f)
-        {
-            Console.WriteLine("More of the same");
-        }
-        else if (_currentPressure < _lastPressure)
-        {
-            Console.WriteLine("Watch out for cooler, rainy weather");
-        }
     }
 }
