@@ -6,7 +6,7 @@ public class Character
 {
     private ICharacterState _characterState = new NormalState();
     public int Health { get; private set; } = 100;
-    public bool HasPowerUp { get; private set; } = false;
+    public bool HasPowerUp { get; private set; }
 
     public void SetState(ICharacterState characterState)
     {
