@@ -1,11 +1,11 @@
-﻿namespace State.Template.DT.States;
+﻿namespace State.Template.States;
 
-public class ConcreteStateB: IState
+public class ConcreteStateB : IState
 {
     public void Handle(Context context)
     {
         Console.WriteLine("ConcreteStateB handles the request");
-        
+
         context.SetState(new ConcreteStateA());
     }
 }
