@@ -7,7 +7,7 @@ public class WhatsappService : IOrderNotifier
         var message = $"Order {order.Id} has been shipped";
         Send(order.CustomerPhone, message);
     }
-    
+
     public void Send(string orderCustomerPhone, string message)
     {
         Console.WriteLine($"Sending Whatsapp message {message} to ${orderCustomerPhone} ");
